Renaming the BaseOperation #run to #call and then we can change from Create.run() to Create.() in the Proc shortcut syntax. Also moving the modules from BaseOperations and StoryOperations to their own files to keep them easier to organize.
