+++ spec/controllers/projects_controller_spec.rb
@@ -29,6 +29,7 @@ describe ProjectsController do
       allow(subject).to receive_messages(:current_user => user)
       allow(user).to receive_messages(:projects => projects)
       allow(user).to receive_message_chain(:projects, :not_archived) { projects }
+      allow(projects).to receive_messages(:pluck => [])
     end
 
     describe "collection actions" do
@@ -62,7 +63,7 @@ describe ProjectsController do
           allow(projects).to receive(:build).with({}) { project }
           allow(project).to receive_messages(:users => users)
           expect(users).to receive(:<<).with(user)
-          allow(ProjectOperations::Create).to receive(:run).with(project, user).and_return(project)
+          allow(ProjectOperations::Create).to receive(:call).with(project, user).and_return(project)
         end
 
         specify do
@@ -83,7 +84,7 @@ describe ProjectsController do
         context "when save fails" do
 
           before do
-            allow(ProjectOperations::Create).to receive(:run).with(project, user).and_return(false)
+            allow(ProjectOperations::Create).to receive(:call).with(project, user).and_return(false)
           end
 
           specify do
@@ -172,7 +173,7 @@ describe ProjectsController do
       describe "#update" do
 
         before do
-          allow(ProjectOperations::Update).to receive(:run).with(project, {}, user).and_return(project)
+          allow(ProjectOperations::Update).to receive(:call).with(project, {}, user).and_return(project)
         end
 
         specify do
@@ -192,7 +193,7 @@ describe ProjectsController do
         context "when update fails" do
 
           before do
-            allow(ProjectOperations::Update).to receive(:run).with(project, {}, user).and_return(false)
+            allow(ProjectOperations::Update).to receive(:call).with(project, {}, user).and_return(false)
           end
 
           specify do
@@ -208,7 +209,7 @@ describe ProjectsController do
       describe "#destroy" do
 
         before do
-          allow(ProjectOperations::Destroy).to receive(:run).with(project, user).and_return(project)
+          allow(ProjectOperations::Destroy).to receive(:call).with(project, user).and_return(project)
         end
 
         specify do
